Reverting commit 0d5d5f95097946e2ed4075b521f7fc4b8cddc72c. Deleted preview related classes in conflict.
