+++ cdap-app-fabric/src/main/java/co/cask/cdap/app/guice/AppFabricServiceRuntimeModule.java
@@ -352,6 +352,14 @@ public final class AppFabricServiceRuntimeModule extends RuntimeModule {
       }
     }
 
+    @Provides
+    @Named(Constants.AppFabric.SERVER_ADDRESS)
+    @SuppressWarnings("unused")
+    public InetAddress providesHostname(CConfiguration cConf) {
+      String address = cConf.get(Constants.AppFabric.SERVER_ADDRESS);
+      return Networks.resolve(address, new InetSocketAddress("localhost", 0).getAddress());
+    }
+
     /**
      * Provides a supplier of quartz scheduler so that initialization of the scheduler can be done after guice
      * injection. It returns a singleton of Scheduler.
