+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/runtime/DataSetsModules.java
@@ -17,6 +17,8 @@
 package co.cask.cdap.data.runtime;
 
 import co.cask.cdap.api.dataset.module.DatasetDefinitionRegistry;
+import co.cask.cdap.common.conf.CConfiguration;
+import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.runtime.RuntimeModule;
 import co.cask.cdap.data2.datafabric.dataset.RemoteDatasetFramework;
 import co.cask.cdap.data2.dataset2.DatasetDefinitionRegistryFactory;
@@ -26,7 +28,9 @@ import co.cask.cdap.data2.dataset2.InMemoryDatasetFramework;
 import co.cask.cdap.data2.metadata.lineage.LineageStore;
 import co.cask.cdap.data2.metadata.lineage.LineageStoreReader;
 import co.cask.cdap.data2.metadata.lineage.LineageStoreWriter;
+import co.cask.cdap.data2.metadata.publisher.KafkaMetadataChangePublisher;
 import co.cask.cdap.data2.metadata.publisher.MetadataChangePublisher;
+import co.cask.cdap.data2.metadata.publisher.NoOpMetadataChangePublisher;
 import co.cask.cdap.data2.metadata.store.DefaultMetadataStore;
 import co.cask.cdap.data2.metadata.store.MetadataStore;
 import co.cask.cdap.data2.metadata.store.NoOpMetadataStore;
@@ -36,8 +40,11 @@ import co.cask.cdap.data2.metadata.writer.LineageWriterDatasetFramework;
 import co.cask.cdap.data2.registry.DefaultUsageRegistry;
 import co.cask.cdap.data2.registry.RuntimeUsageRegistry;
 import co.cask.cdap.data2.registry.UsageRegistry;
+import com.google.inject.Inject;
+import com.google.inject.Injector;
 import com.google.inject.Module;
 import com.google.inject.PrivateModule;
+import com.google.inject.Provider;
 import com.google.inject.Scopes;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
 import com.google.inject.name.Names;
@@ -165,4 +172,24 @@ public class DataSetsModules extends RuntimeModule {
       }
     };
   }
+
+  private static final class MetadataChangePublisherProvider implements Provider<MetadataChangePublisher> {
+    private final Injector injector;
+    private final CConfiguration cConf;
+
+    @Inject
+    MetadataChangePublisherProvider(Injector injector, CConfiguration cConf) {
+      this.injector = injector;
+      this.cConf = cConf;
+    }
+
+    @Override
+    public MetadataChangePublisher get() {
+      if (cConf.getBoolean(Constants.Metadata.UPDATES_PUBLISH_ENABLED)) {
+        return injector.getInstance(KafkaMetadataChangePublisher.class);
+      }
+      return injector.getInstance(NoOpMetadataChangePublisher.class);
+    }
+  }
+
 }
