+++ cdap-data-fabric/src/main/java/co/cask/cdap/data2/dataset2/lib/table/leveldb/LevelDBTableService.java
@@ -61,8 +61,10 @@ public class LevelDBTableService {
 
   private final ConcurrentMap<String, DB> tables = Maps.newConcurrentMap();
 
+  private static final LevelDBTableService SINGLETON = new LevelDBTableService();
+
   public static LevelDBTableService getInstance() {
-    return new LevelDBTableService();
+    return SINGLETON;
   }
 
   /**
