+++ cdap-standalone/src/main/resources/cdap-site.xml
@@ -144,6 +144,7 @@
     </description>
   </property>
 
+
   <!-- Zookeeper Configuration (only used when Zookeeper server is started in Standalone) -->
 
   <property>
@@ -160,14 +161,6 @@
     <description>
       ZooKeeper quorum string; specifies the ZooKeeper host:port; substitute the quorum
       (FQDN1:2181,FQDN2:2181,...) for the components shown here
-    </description>
-  </property>
-
-  <property>
-    <name>enable.alpha.preview</name>
-    <value>false</value>
-    <description>
-      Determines whether to enable preview in standalone
     </description>
   </property>
