+++ cdap-standalone/src/test/java/co/cask/cdap/StandaloneTester.java
@@ -76,7 +76,6 @@ public class StandaloneTester extends ExternalResource {
     cConf.setBoolean(Constants.Explore.START_ON_DEMAND, false);
     cConf.setBoolean(StandaloneMain.DISABLE_UI, true);
     cConf.setBoolean(Constants.Audit.ENABLED, false);
-    cConf.setBoolean(Constants.Preview.ENABLED, false);
 
     for (int i = 0; i < configs.length; i += 2) {
       cConf.set(configs[i].toString(), configs[i + 1].toString());
